+++ ereuse_devicehub/static/js/main.js
@@ -48,12 +48,12 @@
 
   /**
    * Search bar toggle
-   */
   if (select(".search-bar-toggle")) {
     on("click", ".search-bar-toggle", (e) => {
       select(".search-bar").classList.toggle("search-bar-show")
     })
   }
+   */
 
   /**
    * Navbar links active state on scroll