+++ tests/test_basic.py
@@ -117,4 +117,4 @@ def test_api_docs(client: Client):
         'scheme': 'basic',
         'name': 'Authorization',
     }
-    assert len(docs['definitions']) == 132
+    assert len(docs['definitions']) == 134